+++ src/NetworkDynamics.jl
@@ -38,7 +38,7 @@ import SymbolicIndexingInterface as SII
 using SymbolicIndexingInterface: variable_symbols, parameter_symbols
 using StaticArrays: StaticArrays, SVector
 
-export implicit_output
+export implicit_output, ComponentPostprocessing
 include("utils.jl")
 
 export VertexModel, EdgeModel